refactor hooks to their own folder
